Updated files
A couple things were fixed and a calculateStats function was added to
change the players stats according to age, convert status.  A couple
question comments were put above code blocks which need to be resolved.
Still need to fix a bug: it goes into an infinite loop if a non-integer
is input for age or yearsBaptized.